feat(regional): enable transaction log for germany
